version 0.9.25

+++ src/binarize.cpp
@@ -627,7 +627,7 @@ void estimation::estBinarized(int selectedEstimator, int contAttrFrom, int contA
 
 		   binPartition Generator(NoValues) ;
            noPartitions = 0 ;
-		   adjustTables(0,  firstFreeDiscSlot + addedAttr + Mmin(Generator.noPositions(), long(fTree->opt->discretizationSample))) ;
+		   adjustTables(0,  firstFreeDiscSlot + addedAttr + Mmin(Generator.noPositions(), (long long)(fTree->opt->discretizationSample))) ;
            discFrom[estIdx] = firstFreeDiscSlot + addedAttr ;
  		   while (Generator.increment() )
 		   {
